requirements.txt
streamlit
opencv-python
numpy
pillow
streamlit-webrtc
deepface
tensorflow
requests
